requirements.txt
# Core data science
pandas>=2.0.0
numpy>=1.24.0
matplotlib>=3.7.0
seaborn>=0.12.0
scikit-learn>=1.3.0

# ML & Search
sentence-transformers>=2.2.0
faiss-cpu>=1.7.4

# AI
google-generativeai>=0.3.0

# Web UI
gradio>=4.0.0

# Notebook
jupyter>=1.0.0

# Environment
python-dotenv>=1.0.0
